perf: Replace Double with float
We do not need 15 decimal place accuracy
7 is already too much for us

+++ app/src/main/kotlin/com/looker/kenko/data/model/Rating.kt
@@ -18,6 +18,6 @@ import androidx.compose.runtime.Immutable
 
 @Immutable
 @JvmInline
-value class Rating(val value: Double)
+value class Rating(val value: Float)
 
 operator fun Rating.plus(other: Rating) = Rating(value + other.value)
